requirements.txt
Flask
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Werkzeug
uwsgi
psycopg2
